requirements.txt
pandas
gunicorn
dash
dash-ag-grid
dash-bootstrap-components
dash-bootstrap-templates
dash-mantine-components>=0.15.1
dash-iconify
python-dotenv
networkx
prefixed
statsmodels